fundamentos/banco_dados/requirements.txt
mysql-connector==2.2.9
setuptools==75.4.0
